backend/requirements.txt
openai>=1.12.0
tavily-python>=0.3.0
supabase>=2.3.0
python-dotenv>=1.0.0
pydantic>=2.5.0
python-dateutil>=2.8.2
requests>=2.31.0

